requirements-hclass-common.txt
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
firebase-admin==6.0.1